third_party/blink/web_tests/fast/dom/shadow/form-in-shadow-expected.txt

CONSOLE WARNING: line 15: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: line 15: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
PASS obj.hidden is 'hidden'
PASS obj.text is 'text'
PASS obj.checkbox1 is 'on'
PASS obj.checkbox2 is undefined.
PASS obj.range is '50'
PASS obj.textarea is 'textarea'
PASS successfullyParsed is true

TEST COMPLETE
